# Speed up relocation and fix bug