- `relocate_links` was skipped for macOS
- Remove the sequential filter of binaries (very slow)

LLVM before:

```
elf      47.34s
text      0.01s
filter   84.72s
text bin  0.003s
```

LLVM after:

```
elf      13.37s # not sure why this is so flaky...
text      0.01s
text bin  4.81s
```

All in all llvm installs in < 1 min, the bottleneck is decompressing the tarball now.