@gartung 2022-11-22
If I recall correctly I was trying to preserve backwards compatibility.